fixed casting problem, removed min as it caused compiling issues

+++ Port-Scanner/PortScanner.h
@@ -2,11 +2,9 @@
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <string>
-#include <stdint.h>
-#include <mutex>
+#include <cstdint>
 #include <vector>
 #include <algorithm>
-
 #pragma comment(lib, "Ws2_32.lib")
 
 constexpr auto MAX_PORTS = 65535;
@@ -18,10 +16,10 @@ class PortScanner {
 		PortScanner(const std::string& address);
 		
 		void SetAddress(const std::string& address);
-		std::vector<uint32_t> Scan();
+		std::vector<uint16_t> Scan();
 
 	private:
 		std::string address;
 
-		bool CanConnect(uint32_t port);
+		bool CanConnect(uint16_t port);
 };
